// src/app/api/model/ticket-category.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum TicketCategory {
    ACCOUNT_ISSUE = 'ACCOUNT_ISSUE',
    BILLING_PAYMENT = 'BILLING_PAYMENT',
    TECHNICAL_PROBLEM = 'TECHNICAL_PROBLEM',
    FEATURE_REQUEST = 'FEATURE_REQUEST',
    PARTNERSHIP_ISSUE = 'PARTNERSHIP_ISSUE',
    CONTENT_MODERATION = 'CONTENT_MODERATION',
    GENERAL_INQUIRY = 'GENERAL_INQUIRY',
    EARLY_ACCESS_INTEREST = 'EARLY_ACCESS_INTEREST',
    OTHER = 'OTHER'
}